Move shared Cask examples closer to its clients
In Cask specs, we have recently introduced a couple of specialized
`shared_examples` directories. Those new directories are organized
below their respective packages (for instance,
`cask/artifact/shared_examples` and `cask/cmd/shared_examples`).

This commit moves the rest of the shared examples accordingly.

See also:
https://github.com/Homebrew/brew/pull/5247
