+++ Library/Homebrew/test/cask/dsl/shared_examples/base.rb
