+++ Library/Homebrew/test/cask/dsl/shared_examples/staged.rb
